# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Car Dealer API",
    "version": "1.0.0",
    "description": "API documentation for Customer Routes"
  },
  "servers": [
    {
      "url": "http://localhost:4001",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/auth/register/customer": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Register a new customer",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "testcustomer@example.com" },
                  "password": { "type": "string", "example": "password123" },
                  "name": { "type": "string", "example": "Test Customer" },
                  "phone": { "type": "string", "example": "1234567890" }
                },
                "required": ["email", "password", "name", "phone"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Customer registered successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "type": "object" },
                    "accessToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Login",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "example": "testcustomer@example.com" },
                  "password": { "type": "string", "example": "password123" }
                },
                "required": ["email", "password"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": { "type": "object" },
                    "accessToken": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}